Refactor logic into search-existing component
