+++ src/components/SearchExisting.vue
@@ -1,9 +1,33 @@
 <script setup lang="ts">
-defineProps<{ searchMessage: string }>();
-// defineEmits<{}>();
+import { useMessages } from '@/plugins/MessagesPlugin/Messages';
+import { useSearchLinker } from '@/plugins/SearchLinkerPlugin/SearchLinker';
+import { computed } from 'vue';
+import { useStore } from 'vuex';
+
+const messages = useMessages();
+const searchLinker = useSearchLinker();
+const store = useStore();
+const searchUrl = computed( () => {
+	const searchTerm = store.state.lemma;
+	return searchLinker.getSearchUrlForLexeme( searchTerm );
+} );
+const searchMessage = computed( () => messages.get(
+	'wikibaselexeme-newlexeme-search-existing',
+	searchUrl.value,
+) );
+
 </script>
 
 <template>
 	<!-- eslint-disable-next-line vue/no-v-html -->
 	<p class="wbl-snl-intro-text" v-html="searchMessage" />
 </template>
+
+<style lang="scss" scoped>
+@import "@wmde/wikit-tokens/variables";
+@import "@wmde/wikit-vue-components/src/styles/mixins/Typography";
+
+.wbl-snl-intro-text {
+	@include body;
+}
+</style>
